more fixes

+++ services/director-v2/src/simcore_service_director_v2/modules/dynamic_sidecar/api_client/_base.py
@@ -125,7 +125,7 @@ class BaseThinClient:
         base_url: URLTypes | None = None,
         timeout: TimeoutTypes | None = None,
     ) -> None:
-        self.request_timeout: int = request_timeout
+        self.request_timeout: float = request_timeout
 
         client_args: dict[str, Any] = {
             # NOTE: the default httpx pool limit configurations look good
